Show example of private repo credentials
When I followed the instructions in https://www.arthurkoziel.com/setting-up-argocd-with-helm/ my ArgoCD root application stopped working when I added the ArgoCD application to itself to manage (via the "root" app structure). The reason is that my deployment repo is private and requires credentials, and manually entered credentials are lost when ArgoCD tries to sync itself after Helm is removed as the manager.

Additionally, although unrelated, a common scenario when deploying applications that define "mutating webhooks" is that ArgoCD shows the app as out-of-sync when it really is not. The resource.customizations field I added here ignores the persistent caBundle difference.